fix: Players getting warped out of freelance job vehicles

+++ PlayerSystems/Jobs/BusDriver/BusDriverUtils.cs
@@ -10,7 +10,7 @@ namespace CloudRP.PlayerSystems.Jobs.BusDriver
         public string depoName { get; set; }
         public Vector3 depoStartPosition { get; set; }
         public int depoId { get; set; }
-        public double busStartRotation { get; set; }
+        public float busStartRotation { get; set; }
         public Vector3 busStartPosition { get; set; }
     }
